requirements.txt
Python 3.6 or higher
Streamlit
OpenCV
Pillow (PIL)